refactor group
